Refactor: Basic folder structure migration to new NextJS app router
