+++ src/app/layout.tsx
@@ -1,0 +1,16 @@
+import "~/styles/globals.css"
+
+export const metadata = {
+  title: "ludolab | Lista de juegos",
+  description: "Conoce juegos y dinámicas de grupo para hacer con amigos",
+}
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang="es">
+      <body>
+        {children}
+      </body>
+    </html>
+  )
+}
