+++ src/components/ApolloClientProvider.tsx
@@ -1,16 +1,12 @@
+"use client"
 import { ApolloClient, ApolloProvider, InMemoryCache } from "@apollo/client"
+import React from "react"
 
 const client = new ApolloClient({
   uri: process.env.NEXT_PUBLIC_GRAPHQL_ENDPOINT,
   cache: new InMemoryCache(),
 })
 
-type GraphQLClientProviderProps = {
-  children: JSX.Element
-}
-
-const GraphQLClientProvider = ({ children }: GraphQLClientProviderProps) => {
+export default function ApolloClientProvider({ children }: { children: React.ReactNode }) {
   return <ApolloProvider client={client}>{children}</ApolloProvider>
 }
-
-export default GraphQLClientProvider
