+++ src/hooks/usePage.ts
@@ -1,24 +1,23 @@
-import { NextRouter } from "next/router"
 import { useEffect, useState } from "react"
 import { PageState } from "~/types"
 
-const usePage = (router: NextRouter): PageState => {
+const usePage = (searchParams: URLSearchParams): PageState => {
   const itemsPerPage = Number(process.env.NEXT_PUBLIC_GAMES_PER_PAGE) || 6
   const [page, setPage] = useState(1)
   const [itemsToSkip, setItemsToSkip] = useState(0)
 
   useEffect(() => {
-    const queryPage = Number(router.query.page)
+    const queryPage = Number(searchParams.get("page"))
 
-    if (isNaN(queryPage)) {
+    if (isNaN(queryPage) || queryPage < 1) {
       setPage(1)
     } else {
       setPage(queryPage)
     }
 
-    const itemsToSkip = itemsPerPage * (page - 1)
+    const itemsToSkip = itemsPerPage * ((isNaN(queryPage) || queryPage < 1 ? 1 : queryPage) - 1)
     setItemsToSkip(itemsToSkip)
-  }, [router, page, itemsPerPage])
+  }, [searchParams, itemsPerPage])
 
   return { page, itemsPerPage, itemsToSkip }
 }
