+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -16,16 +20,44 @@
     "incremental": true,
     "baseUrl": "src",
     "paths": {
-      "~/graphql/*": ["graphql/*"],
-      "~/lib/*": ["lib/*"],
-      "~/hooks/*": ["hooks/*"],
-      "~/components/*": ["components/*"],
-      "~/public/*": ["public/*"],
-      "~/styles/*": ["styles/*"],
-      "~/utils/*": ["utils/*"],
-      "~/types": ["types.d.ts"],
-    }
+      "~/graphql/*": [
+        "graphql/*"
+      ],
+      "~/lib/*": [
+        "lib/*"
+      ],
+      "~/hooks/*": [
+        "hooks/*"
+      ],
+      "~/components/*": [
+        "components/*"
+      ],
+      "~/public/*": [
+        "public/*"
+      ],
+      "~/styles/*": [
+        "styles/*"
+      ],
+      "~/utils/*": [
+        "utils/*"
+      ],
+      "~/types": [
+        "types.d.ts"
+      ]
+    },
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
-  "exclude": ["node_modules"]
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
